multi-agent-lab-example/requirements.txt
langchain-community==0.3.20
wikipedia==1.4.0
google-cloud-bigquery
langchain-google-vertexai
langchain
1